% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_obscov.r
\name{get_probzero}
\alias{get_probzero}
\title{Get probability of zero bycatch given effort, bycatch rate, and dispersion}
\usage{
get_probzero(n, bpue, d)
}
\arguments{
\item{n}{Integer vector. Observed effort levels (in terms of sets/hauls) for
which to calculate probability of zero bycatch.}

\item{bpue}{Numeric greater than zero. Bycatch per unit effort.}

\item{d}{Numeric >= 1. Negative binomial dispersion parameter. The dispersion
parameter corresponds to the variance-to-mean ratio of set-level bycatch, 
so d=1 corresponds to Poisson-distributed bycatch, and d>1 corresponds to
overdispersed bycatch.}
}
\value{
Vector of same length as n with probabilities of zero bycatch.

Returned invisibly
}
\description{
\code{get_probzero} returns probability of zero bycatch in a specified number 
of sets/hauls, given bycatch per unit effort and negative binomial dispersion 
parameter. Calculated from the probability density at zero of the
corresponding Poisson or negative binomial distribution.
}
