% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cv.r
\name{plot_cv}
\alias{plot_cv}
\title{Plot bycatch estimation CV vs. observer coverage}
\usage{
plot_cv(te, bpue, d = 2, targetcv = 0.3, showplot = TRUE, silent = FALSE, ...)
}
\arguments{
\item{te}{an integer greater than 1. Total effort in fishery (e.g., trips 
or sets).}

\item{bpue}{a positive number. Bycatch per unit effort.}

\item{d}{a number greater than or equal to 1. Dispersion index. The dispersion 
index corresponds to the variance-to-mean ratio of effort-unit-level bycatch, 
so \code{d = 1} corresponds to Poisson-distributed bycatch, and \code{d > 1} 
to overdispersed bycatch.}

\item{targetcv}{a non-negative number less than 1. Target CV (as a proportion). 
If set to 0, no corresponding minimum observer coverage will be 
highlighted or returned.}

\item{showplot}{logical. If \code{FALSE}, plotting is suppressed.}

\item{silent}{logical. If \code{TRUE}, print output to terminal is suppressed.}

\item{...}{additional arguments for compatibility with Shiny.}
}
\value{
If \code{targetcv} is non-zero, a list with one component:
  \item{targetoc}{minimum observer coverage in terms of percentage.}

Returned invisibly.
}
\description{
\code{plot_cv} plots projected bycatch estimation CVs vs observer 
coverage, and returns minimum observer coverage needed to achieve 
user-specified target CV and percentile.
}
\details{
\strong{Caveat:} \code{plot_cv} assumes that (1) observer coverage is 
representative, (2) bycatch (\code{bpue}) is in terms of individuals (not 
weight) per unit effort, and (3) the specified dispersion index reflects 
the highest level of any hierarchical variance (e.g., using dispersion index 
at trip level if greater than that at set level). Violating these assumptions 
will likely result in negatively biased projections of bycatch estimation CV 
for a given level of observer coverage. More conservative projections can be 
obtained by using a higher dispersion index \code{d}. Users may want to 
explore uncertainty in dispersion index and in bycatch per unit effort by 
varying those inputs.
}
