% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_obscov.r
\name{plot_cv_obscov}
\alias{plot_cv_obscov}
\title{Plot CV vs. observer coverage}
\usage{
plot_cv_obscov(simlist = simlist, targetcv = 30, q = 0.8)
}
\arguments{
\item{simlist}{List output from sim_cv_obscov.}

\item{targetcv}{Numeric, 0 < targetcv <=100. Target CV (as percentage). 
If 0, no corresponding minimum observer coverage will be highlighted.}

\item{q}{Numeric, 0 < q <=0.95. Desired probability (as a proportion) of 
achieving at least target CV or lower.}
}
\value{
A list with minimum observer coverage in terms of percentage ($pobscov) 
  and effort ($nobsets) corresponding to user specifications.

Returned invisibly.
}
\description{
\code{plot_cv_obscov} plots CV of bycatch estimates vs observer coverage for
user-specified percentile (i.e., probability of achieving CV) and several 
default percentiles, and prints minimum observer coverage needed to achieve 
user targets. 
  
\strong{CAVEAT:} Since \code{sim_cv_obscov} assumes representative observer coverage 
and no hierarchical sources of variance (e.g., vessel- or trip-level variation), 
bycatch estimation CV for a given level of observer coverage is likely to be 
biased low relative to the real world.
}
