% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_obscov.r
\name{plot_cvsim_samplesize}
\alias{plot_cvsim_samplesize}
\title{Plot sample size for CV estimates vs observer coverage}
\usage{
plot_cvsim_samplesize(simlist = simlist)
}
\arguments{
\item{simlist}{List output from sim_cv_obscov.}
}
\value{
None
}
\description{
\code{plot_cvsim_samplesize} plots sample size (simulations with positive
observed bycatch) vs observer coverage level, along with probability of 
observing zero bycatch based on effort and the probability density at zero
given bycatch rate and negative binomial dispersion.
}
\details{
Note that the predicted probability of observing zero bycatch assumes 
representative observer coverage and no hierarchical sources of variance 
(e.g., vessel- or trip-level variation), so probability of observing zero 
bycatch at a given level of observer coverage is likely to be biased low 
relative to the real world.
}
