% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_obscov.r
\name{plot_probzeroobs}
\alias{plot_probzeroobs}
\title{Plot probability of observing zero bycatch vs observer coverage}
\usage{
plot_probzeroobs(simdat = simdat)
}
\arguments{
\item{simdat}{Tibble output from sim_obscov_cv.}
}
\value{
None
}
\description{
Plot probability of observing zero bycatch vs observer coverage
}
