% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_obscov.r
\name{plot_samplesize_cvsim}
\alias{plot_samplesize_cvsim}
\title{Plot sample size for CV estimates vs observer coverage}
\usage{
plot_samplesize_cvsim(simlist = simlist)
}
\arguments{
\item{simlist}{List output from sim_obscov_cv.}
}
\value{
None
}
\description{
\code{plot_samplesize_cvsim} plots sample size (simulations with positive
observed bycatch) vs observer coverage level, along with probability of 
observing zero bycatch based on effort and the probability density at zero
given bycatch rate and negative binomial dispersion.
}
