% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_obscov.r
\name{probzero}
\alias{probzero}
\title{Get probability of zero bycatch in one unit of effort}
\usage{
probzero(bpue, d)
}
\arguments{
\item{bpue}{a positive number. Bycatch per unit effort.}

\item{d}{a number greater than or equal to 1. Negative binomial dispersion 
index. The dispersion index corresponds to the variance-to-mean 
ratio of set-level bycatch, so \eqn{d = 1} corresponds to Poisson-distributed 
bycatch, and \eqn{d > 1} corresponds to overdispersed bycatch.}
}
\value{
Scalar probability of zero bycatch.

Returned invisibly
}
\description{
\code{probzero} returns probability of zero bycatch in a one set/haul, given
bycatch per unit effort and negative binomial dispersion index.
}
\details{
Calculated from the probability density at zero of the corresponding Poisson
(\eqn{d = 1}) or negative binomial (\eqn{d < 1}) distribution.

\strong{Caveat:} \code{probzero} assumes representative observer coverage 
and no hierarchical sources of variance (e.g., vessel- or trip-level variation). 
Violating these assumptions will likely result in negatively biased projections 
of the probability of observing zero bycatch at a given level of observer coverage. 
More conservative projections can be obtained by using higher-level units of effort 
(e.g., \code{bpue} as mean bycatch per trip instead of bycatch per set/haul, and 
\code{n} as number of trips instead of number of sets/hauls).
}
