% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_cv_obscov.r
\name{sim_cv_obscov}
\alias{sim_cv_obscov}
\title{Simulate CV response to observer coverage}
\usage{
sim_cv_obscov(te, bpue, d = 2, nsim = 1000, ...)
}
\arguments{
\item{te}{an integer greater than one. Total effort in fishery (e.g., trips 
or sets).}

\item{bpue}{a positive number. Bycatch per unit effort (with bycatch in terms 
of number of individuals).}

\item{d}{a number greater than or equal to 1. Dispersion index. The dispersion 
index corresponds to the variance-to-mean ratio of effort-unit-level bycatch, 
so \code{d = 1} corresponds to Poisson-distributed bycatch, and \code{d > 1} 
to overdispersed bycatch.}

\item{nsim}{a positive integer. Number of simulations to run.}

\item{...}{additional arguments for compatibility with Shiny.}
}
\value{
A list with components:
  \item{simsum}{a tibble with one row per observer coverage level and the 
  following fields: simulated proportion observer coverage (\code{pobs}), 
  number of observed trips/sets (\code{nobs}), and bycatch estimation CV 
  (\code{cvsim}).} 
  \item{simdat}{a tibble with one row per simulation and the following fields: 
  simulated proportion observer coverage (\code{pobs}), number of observed 
  effort units (\code{nobs}), true (realized) bycatch per unit effort (\code{tbpue}), 
  observed bycatch per unit effort (\code{obpue}), and error of observed bycatch 
  per unit effort (\code{oberr} = \code{obpue} - \code{tbpue}).}
  \item{bpue}{nominal bycatch per unit effort used in the simulations.}
  \item{d}{dispersion index used in the simulations.}
}
\description{
\code{sim_cv_obscov} simulates bycatch estimation CVs for a range 
of observer coverage levels, given bycatch rate (as number of individuals per 
unit effort), dispersion index of bycatch, and total fishery effort. 
The resulting object is required by \code{plot_cv_obscov} for visualization.
}
\details{
\code{sim_cv_obscov} runs \code{nsim} simulations per level of observer 
coverage, from the larger of 0.1\% or one effort unit to 100\%. Simulated 
bycatch estimates are calculated as mean observed bycatch per unit effort. 
CV at each observer coverage level is calculated as the square root of mean 
square estimation error divided by "true" bycatch (\code{bpue}).

Warning: Large total effort (>100K trips/sets) may require several minutes 
of execution time. Increasing \code{nsim} from the default of 1000 will 
also increase execution time. 

\strong{Caveat:} \code{sim_cv_obscov} assumes that (1) observer coverage is 
representative, (2) bycatch (\code{bpue}) is in terms of individuals (not 
weight) per unit effort, and (3) the specified dispersion index reflects 
the highest level of any hierarchical variance (e.g., using dispersion index 
at trip level if greater than that at set level). Violating these assumptions 
will likely result in negatively biased projections of bycatch estimation CV 
for a given level of observer coverage. More conservative projections can be 
obtained by using a higher dispersion index \code{d}. Users may want to 
explore uncertainty in dispersion index and in bycatch per unit effort by 
varying those inputs.
}
