% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_obscov.r
\name{get_probzero}
\alias{get_probzero}
\title{Get probability of zero bycatch given effort, bycatch rate, and dispersion}
\usage{
get_probzero(n, bpue, d)
}
\arguments{
\item{n}{a vector of positive integers. Observed effort levels (in terms of 
sets/hauls) for which to calculate probability of zero bycatch.}

\item{bpue}{a positive number. Bycatch per unit effort.}

\item{d}{a number greater than or equal to 1. Negative binomial dispersion 
parameter. The dispersion parameter corresponds to the variance-to-mean 
ratio of set-level bycatch, so \eqn{d = 1} corresponds to Poisson-distributed 
bycatch, and \eqn{d > 1} corresponds to overdispersed bycatch.}
}
\value{
Vector of same length as \code{n} with probabilities of zero bycatch.

Returned invisibly
}
\description{
\code{get_probzero} returns probability of zero bycatch in a specified number 
of sets/hauls, given bycatch per unit effort and negative binomial dispersion 
parameter.
}
\details{
Calculated from the probability density at zero of the corresponding Poisson
(\eqn{d = 1}) or negative binomial (\eqn{d < 1}) distribution.

\strong{Caveat:} \code{get_probzero} assumes representative observer coverage 
and no hierarchical sources of variance (e.g., vessel- or trip-level variation), 
so probability of observing zero bycatch at a given level of observer coverage 
is likely to be biased low relative to the real world. More conservative 
estimates can be obtained by using higher-level bycatch and effort information 
(e.g., \code{bpue} as mean bycatch per trip instead of bycatch per set/haul, and 
\code{n} as number of trips instead of number of sets/hauls).
}
