% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cv_obscov.r
\name{plot_cv_obscov}
\alias{plot_cv_obscov}
\title{Plot bycatch estimation CV vs. observer coverage}
\usage{
plot_cv_obscov(simlist = simlist, targetcv = 0.3, showplot = TRUE,
  silent = FALSE, ...)
}
\arguments{
\item{simlist}{list output from \code{sim_cv_obscov}.}

\item{targetcv}{a non-negative number less than 1. Target CV (as a proportion). 
If set to 0, no corresponding minimum observer coverage will be 
highlighted or returned.}

\item{showplot}{logical. If \code{FALSE}, plotting is suppressed.}

\item{silent}{logical. If \code{TRUE}, print output to terminal is suppressed.}

\item{...}{additional arguments for compatibility with Shiny.}
}
\value{
If \code{targetcv} is non-zero, a list with one component:
  \item{targetoc}{minimum observer coverage in terms of percentage.}

Returned invisibly.
}
\description{
\code{plot_cv_obscov} plots projected bycatch estimation CVs vs observer 
coverage, and returns minimum observer coverage needed to achieve 
user-specified target CV and percentile.
}
\details{
\strong{Caveat:} \code{plot_cv_obscov} assumes that (1) observer coverage is 
representative, (2) bycatch specified for \code{sim_cv_obscov}is in terms of 
individuals (not weight) per unit effort, and (3) the dispersion index 
specified for \code{sim_cv_obscov} reflects the highest level of any 
hierarchical variance (e.g., using dispersion index at trip level if greater 
than that at set level). Violating these assumptions will likely result in 
negatively biased projections of bycatch estimation CV for a given level of 
observer coverage. Users may want to explore uncertainty in dispersion index 
and in bycatch per unit effort by varying those inputs. See documentation for 
\code{sim_cv_obscov} for additional details.
}
