% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_obscov.r
\name{plot_cv_obscov}
\alias{plot_cv_obscov}
\title{Plot CV vs. observer coverage}
\usage{
plot_cv_obscov(simlist = simlist, targetcv = 30, q = 0.8)
}
\arguments{
\item{simlist}{list output from \code{sim_cv_obscov}.}

\item{targetcv}{a non-negative number less than or equal to 100. Target CV 
(as percentage). If \eqn{targetcv = 0}, no corresponding minimum observer 
coverage will be highlighted.}

\item{q}{a positive number less than or equal to 0.95. Custom percentile to 
be plotted, desired probability (as a proportion) of achieving the target 
CV or lower. Ignored if \eqn{targetcv = 0}.}
}
\value{
A list with components:
  \item{pobscov}{minimum observer coverage in terms of percentage.} 
  \item{nobsets}{corresponding observed effort.}

Returned invisibly.
}
\description{
\code{plot_cv_obscov} plots CV of bycatch estimates vs observer coverage for
user-specified percentile (i.e., probability of achieving CV) and several 
default percentiles, and returns minimum observer coverage needed to achieve 
user-specified target CV and percentile.
}
\details{
\strong{Caveat:} Since \code{sim_cv_obscov} assumes representative observer 
coverage and no hierarchical sources of variance (e.g., vessel- or trip-level 
variation), bycatch estimation CV for a given level of observer coverage is 
likely to be biased low relative to the real world. See documentation for
\code{sim_obs_cov} for additional details.
}
