% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cv_obscov.r
\name{plot_cv_obscov}
\alias{plot_cv_obscov}
\title{Plot bycatch estimation CV vs. observer coverage}
\usage{
plot_cv_obscov(
  simlist = simlist,
  targetcv = 0.3,
  showplot = TRUE,
  silent = FALSE
)
}
\arguments{
\item{simlist}{list output from \code{sim_cv_obscov}.}

\item{targetcv}{a non-negative number less than 1. Target CV 
(as a proportion). If \eqn{targetcv = 0}, no corresponding minimum observer 
coverage will be highlighted.}

\item{showplot}{logical. If plot = FALSE, plotting is suppressed.}

\item{silent}{logical. If silent = TRUE, print output to terminal is suppressed.}
}
\value{
A list with components:
  \item{pobs}{minimum observer coverage in terms of percentage.} 
  \item{nobs}{corresponding observed effort.}

Returned invisibly.
}
\description{
\code{plot_cv_obscov} plots projected bycatch estimation CVs vs observer 
coverage, and returns minimum observer coverage needed to achieve 
user-specified target CV and percentile.
}
\details{
\strong{Caveat:} \code{sim_cv_obscov} assumes representative observer 
coverage and no hierarchical sources of variance (e.g., vessel- or trip-level 
variation). Violating these assumptions will likely result in negatively biased 
projections of bycatch estimation CV for a given level of observer coverage. 
See documentation for \code{sim_obs_cov} for additional details.
}
