% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_obscov.r
\name{plot_obscov_cv}
\alias{plot_obscov_cv}
\title{Plot CV vs. observer coverage}
\usage{
plot_obscov_cv(simdat = simdat, targetcv = 30, q = 0.8)
}
\arguments{
\item{simdat}{Tibble output from sim_obscov_cv.}

\item{targetcv}{Numeric, 0 < targetcv <=100. Target CV (as percentage).}

\item{q}{Numeric, 0 < q <=0.95. Desired probability (as a proportion) of 
achieving at least target CV or lower.}
}
\value{
Numeric scalar. Estimate of minimum observer coverage (percentage) 
  required to meet target CV.
}
\description{
Plot CV vs. observer coverage
}
