% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_obscov.r
\name{plot_probposobs}
\alias{plot_probposobs}
\title{Plot probability of positive observed bycatch vs observer coverage}
\usage{
plot_probposobs(te, bpue, d, target.ppos = 80)
}
\arguments{
\item{te}{Integer scalar greater than 10. Total effort in fishery (sets/hauls).}

\item{bpue}{Numeric greater than zero. Bycatch per unit effort.}

\item{d}{Numeric >= 1. Negative binomial dispersion parameter. The dispersion
parameter corresponds to the variance-to-mean ratio of set-level bycatch, 
so d=1 corresponds to Poisson-distributed bycatch, and d>1 corresponds to
overdispersed bycatch.}

\item{target.ppos}{Numeric, 0 < target.ppos <=100. Target probability of
positive observed bycatch (as percentage), given positive bycatch in total 
effort. If 0, no corresponding minimum observer coverage will be highlighted.}
}
\value{
A list with minimum observer coverage in terms of percentage ($pobscov) 
  and effort ($nobsets) corresponding to user specifications.

Returned invisibly.
}
\description{
\code{plot_probposobs} plots probability of observing at least one bycatch 
  event vs observer coverage, given total effort in sets/hauls, bycatch per 
  unit effort, and negative binomial dispersion parameter. Probabilities are
  based on the probability density function for the corresponding Poisson or 
  negative binomial distribution.
}
