% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_obscov.r
\name{plot_probposobs}
\alias{plot_probposobs}
\title{Plot probability of positive bycatch vs observer coverage}
\usage{
plot_probposobs(te, bpue, d = 2, targetppos = 95, showplot = TRUE,
  silent = FALSE)
}
\arguments{
\item{te}{an integer greater than 1. Total effort in fishery (trips/sets).}

\item{bpue}{a positive number. Bycatch per unit effort.}

\item{d}{a number greater than or equal to 1. Dispersion 
index. The dispersion index corresponds to the variance-to-mean 
ratio of effort-unit-level bycatch, so \eqn{d = 1} corresponds to Poisson-
distributed bycatch, and \eqn{d > 1} corresponds to overdispersed bycatch.}

\item{targetppos}{a non-negative number less than or equal to 100. Target 
probability of positive observed bycatch (as percentage), given positive 
bycatch in total effort. If 0, no corresponding minimum observer coverage 
will be highlighted.}

\item{showplot}{logical. If plot = FALSE, plotting is suppressed.}

\item{silent}{logical. If silent = TRUE, print output to terminal is suppressed.}
}
\value{
A list with components:
  \item{pobs}{minimum observer coverage in terms of percentage.} 
  \item{nobs}{corresponding observed effort.}
  \item{ppos.te}{probability of any bycatch occurring in total effort}

Returned invisibly.
}
\description{
\code{plot_probposobs} plots (1) probability of observing at least one bycatch
  event vs observer coverage and (2) probability of any bycatch occurring in 
  total effort, given total effort in trips/sets, bycatch per unit effort, and 
  dispersion index.
}
\details{
Probabilities are based on the probability density function for the 
corresponding Poisson or negative binomial distribution.

The probability that any bycatch occurs in the given total effort is shown
by the horizontal black dotted line. The conditional probability of observing 
any bycatch if it occurs is shown by the solid black line.  The product of 
these first two probabilities gives the absolute probability of observing any
bycatch (dashed black line).The minimum observer coverage to achieve the target 
obability of observing bycatch if it occurs (x-axis value of red star) is 
where the conditional bycatch detection probability (solid black line) 
intersects with the target probability (red dash-dot line).

\strong{Caveat:} \code{plot_probposobs} assumes representative observer coverage 
and no hierarchical sources of variance (e.g., vessel- or trip-level variation). 
Violating these assumptions will likely result in positively biased projections 
of the probability of observing bycatch at a given level of observer coverage. 
More conservative projections can be obtained by using higher-level units of effort 
(e.g., \code{bpue} as mean bycatch per trip instead of bycatch per trip/set, and 
\code{te} as number of trips instead of number of trips/sets).
}
