% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_obscov.r
\name{plot_probposobs}
\alias{plot_probposobs}
\title{Plot probability of positive observed bycatch vs observer coverage}
\usage{
plot_probposobs(te, bpue, d = 2, target.ppos = 80)
}
\arguments{
\item{te}{an integer greater than 1. Total effort in fishery (sets/hauls).}

\item{bpue}{a positive number. Bycatch per unit effort.}

\item{d}{a number greater than or equal to 1. Negative binomial dispersion 
parameter. The dispersion parameter corresponds to the variance-to-mean 
ratio of set-level bycatch, so \eqn{d = 1} corresponds to Poisson-distributed 
bycatch, and \eqn{d > 1} corresponds to overdispersed bycatch.}

\item{target.ppos}{a non-negative number less than or equal to 100. Target 
probability of positive observed bycatch (as percentage), given positive 
bycatch in total effort. If 0, no corresponding minimum observer coverage 
will be highlighted.}
}
\value{
A list with components:
  \item{pobscov}{minimum observer coverage in terms of percentage.} 
  \item{nobsets}{corresponding observed effort.}

Returned invisibly.
}
\description{
\code{plot_probposobs} plots probability of observing at least one bycatch 
  event vs observer coverage, given total effort in sets/hauls, bycatch per 
  unit effort, and negative binomial dispersion parameter.
}
\details{
Probabilities are based on the probability density function for the 
corresponding Poisson or negative binomial distribution.

The probability that any bycatch occurs in the given total effort is shown
by the horizontal black dotted line. The conditional probability of observing 
any bycatch if it occurs is shown by the solid black line.  The product of 
these first two probabilities gives the absolute probability of observing any
bycatch (dashed black line).The minimum observer coverage to achieve the target 
obability of observing bycatch if it occurs (x-axis value of red star) is 
where the conditional bycatch detection probability (solid black line) 
intersects with the target probability (red dash-dot line).

Note that unlike \code{plot_cv_obscov}, \code{plot_probposobs} is designed 
as a one-step tool, and does not take output from user calls to 
\code{get_probzero}. 
  
\strong{Caveat:} \code{plot_probposobs} assumes representative observer coverage 
and no hierarchical sources of variance (e.g., vessel- or trip-level variation). 
Violating these assumptions will likely result in positively biased projections 
of the probability of observing bycatch at a given level of observer coverage. 
More conservative projections can be obtained by using higher-level units of effort 
(e.g., \code{bpue} as mean bycatch per trip instead of bycatch per set/haul, and 
\code{te} as number of trips instead of number of sets/hauls).
}
