% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_obscov.r
\name{plot_samplesize_cvsim}
\alias{plot_samplesize_cvsim}
\title{Plot sample size for CV estimates vs observer coverage}
\usage{
plot_samplesize_cvsim(simdat = simdat, bpue, d)
}
\arguments{
\item{simdat}{Tibble output from sim_obscov_cv.}

\item{bpue}{Numeric greater than zero. Bycatch per unit effort.}

\item{d}{Numeric >= 1. Negative binomial dispersion parameter. The dispersion
parameter corresponds to the variance-to-mean ratio of set-level bycatch, 
so d=1 corresponds to Poisson-distributed bycatch, and d>1 corresponds to
overdispersed bycatch.}
}
\value{
None
}
\description{
\code{plot_samplesize_cvsim} plots sample size (simulations with positive
observed bycatch) vs observer coverage level, along with probability of 
observing zero bycatch based on effort and the probability density at zero
given bycatch rate and negative binomial dispersion.
}
