% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_uclnegobs.r
\name{plot_uclnegobs}
\alias{plot_uclnegobs}
\title{Plot upper confidence limit of total bycatch given none observed}
\usage{
plot_uclnegobs(te, d = 2, cl = 95, targetucl = 0, fixedoc = 0,
  ymax = 100, showplot = TRUE, silent = FALSE, ...)
}
\arguments{
\item{te}{an integer greater than 1. Total effort in fishery (e.g., trips 
or sets).}

\item{d}{a number greater than or equal to 1. Dispersion index. The dispersion 
index corresponds to the variance-to-mean ratio of effort-unit-level bycatch, 
so \code{d = 1} corresponds to Poisson-distributed bycatch, and \code{d > 1} 
to overdispersed bycatch.}

\item{cl}{a non-negative number less than or equal to 100. Confidence level
for upper confidence limit of total bycatch (as percentage), given no bycatch 
observed.}

\item{targetucl}{a non-negative number. Target maximum upper confidence 
limit for total bycatch given zero bycatch observed. If set to 0, no 
corresponding minimum observer coverage will be highlighted or returned.}

\item{fixedoc}{a non-negative number between 0 and 100. Percent observer coverage 
for which to return ucl value.}

\item{ymax}{a positive number. Upper limit for y-axis of plot.}

\item{showplot}{logical. If \code{FALSE}, plotting is suppressed.}

\item{silent}{logical. If \code{TRUE}, print output to terminal is suppressed.}

\item{...}{additional arguments for compatibility with Shiny.}
}
\value{
A list with components:
  \item{ucldat}{a tibble with the following fields for each coverage level included: 
  number of observed effort units (\code{nobs}), 
  proportion observer coverage (\code{pobs}), 
  upper confidence limit of total bycatch given none observed (\code{ucl}),
  and finite population correction (\code{fpc}) used in calculating \code{ucl}.}
  \item{targetucl}{specified target maximum upper confidence limit of bycatch.}
  \item{targetoc}{minimum observer coverage (as percentage) for which upper 
  confidence limit of bycatch is \code{targetucl} when none observed.}
  \item{fixedoc}{specified percentage observer coverage for which upper 
  confidence limit of bycatch is returned.}
  \item{fixednoc}{observer coverage (as effort) corresponding to \code{fixedoc}.}
  \item{fixedoc.ucl}{upper confidence limit of total bycatch corresponding 
  to zero bycatch observed in \code{fixedoc} coverage.}
  \item{te}{specified total effort.}
  \item{d}{specified dispersion index.}
  \item{cl}{specified confidence level.}

Returned invisibly.
}
\description{
\code{plot_uclnegobs} plots upper confidence limit of total bycatch vs 
  observer coverage when no bycatch is observed, given total fishery effort, 
  dispersion index, and confidence level. The function returns (1) minimum 
  observer coverage needed to fall within user-specified upper confidence 
  limit for bycatch when none was observed, and/or (2) the upper confidence 
  limit for bycatch given specified observer coverage and no observed bycatch.
}
\details{
Upper confidence limits are based on the probability density function for 
the corresponding Poisson or negative binomial distribution. Upper confidence 
limits based on \code{d}+/-1 (as allowed by specification of d) are also plotted. 
If \code{fixedoc} specified, corresponding upper confidence limit is provided 
in printed output and returned object, but not in plot.

\strong{Caveat:} \code{plot_uclnegobs} assumes that (1) observer coverage is 
representative, (2) bycatch is in terms of individuals (not weight) per unit 
effort,and (3) the specified dispersion index reflects the highest level of 
any hierarchical variance (e.g., using dispersion index at trip level if 
greater than that at set level). Violating these assumptions will likely 
result in negatively biased projections of the upper confidence limit of total 
bycatch given zero observed. More conservative projections can be obtained by 
using a higher dispersion index \code{d}. Users may want to explore 
uncertainty in dispersion index and in bycatch per unit effort by varying 
those inputs.
}
