% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_shiny.r
\name{run_shiny}
\alias{run_shiny}
\title{Run Shiny interface with ObsCovgTools}
\usage{
run_shiny()
}
\value{
None
}
\description{
\code{run_shiny} runs a shiny application for the main functions in ObsCovgTools.
}
\details{
Note: Estimated run times in Bycatch Estimation CV tab only apply to execution
on shinyapps.io server (see README)
}
